refactor: create PackagesTestMixin, move some helper functions to it
